DT-32 Tests, banner.txt, profile configs added

+++ src/main/resources/db/migration/V5__create_appointment_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS appointments
 (
-    id INT AUTO_INCREMENT PRIMARY KEY,
+    id BIGINT AUTO_INCREMENT PRIMARY KEY,
     status VARCHAR(20) NOT NULL,
     CHECK  (status IN ('CANCELED', 'CONFIRMED', 'AWAITING')),
     date_and_time TIMESTAMP NOT NULL,
-    patient_id INT,
+    patient_id BIGINT,
     FOREIGN KEY (patient_id) REFERENCES patients(id)
 )
